Move some player controller to a dedicated package.
